move the five bulk operators into mortie/batch.py (issue #170)
