+++ mortie/geometry.py
@@ -406,176 +406,6 @@ def from_wkb(data, order=18, moc=False, normalize=True,
     return _cover_parts(kind, parts, order, moc, normalize, tolerance, max_cells)
 
 
-def from_wkbs(blobs, order=18, tolerance=None, max_cells=None, normalize=True):
-    """Cover many WKB blobs with one call -- ragged MOCs out, no backend.
-
-    The batch sibling of :func:`from_wkb` (issue #157) and the plural twin its
-    name marks: **one MOC per input blob** (many→many), against the many→one
-    union :func:`from_wkb` performs over the rings *inside* one blob.  The
-    whole column crosses the Python/Rust boundary once, and Rust parses and
-    covers the blobs in parallel with the GIL released — so the per-call fixed
-    cost that dominates a Python loop over half a million footprints is paid
-    once.  Result ``i`` is byte-identical to
-    ``from_wkb(blobs[i], order=order, moc=True, ...)``.
-
-    Memory: a chunk ends at 2048 blobs **or 64 MiB, whichever comes first**,
-    and peak is the returned ``values`` array plus **one chunk of copied input
-    bytes** (the copy is mandatory — a Python ``bytes`` buffer is GIL-bound and
-    cannot cross into the parallel region) plus one chunk of in-flight covers.
-    Neither the whole column's bytes nor every blob's cover is ever resident at
-    once, and that holds **for every input spelling and every blob size**:
-    non-``bytes`` entries are coerced inside the chunk, so their copy dies with
-    it, and the byte budget stops 2048 fat geometries from making "one chunk"
-    mean gigabytes.  Measured on the 555,867-blob ATL03 v007 corpus (276.7 MiB
-    of WKB, 167.3 MiB of result, order 6), peak growth over the resident
-    column:
-
-    ==========================  ==========  =========
-    input spelling              peak        × result
-    ==========================  ==========  =========
-    ``list[bytes]``             178.7 MiB   1.07
-    numpy object array          179.4 MiB   1.07
-    hex ``str``                 178.9 MiB   1.07
-    ``bytearray``               178.9 MiB   1.07
-    ``memoryview``              179.4 MiB   1.07
-    ``uint8`` array             221.2 MiB   1.32
-    arrow buffer slices         179.3 MiB   1.07
-    ==========================  ==========  =========
-
-    On the fat end, 3,000 Antarctic-basin blobs (1.25 MiB each, a 3.7 GiB
-    column) peak at 610-634 MiB, against 3,120 MiB when the chunk was bounded
-    by blob count alone — and most of what is left is the in-flight cover
-    work, not the copy.
-
-    Parameters
-    ----------
-    blobs : sequence
-        One WKB/EWKB geometry per entry.  Each entry takes exactly what
-        :func:`from_wkb` takes — ``bytes``, a hex ``str``, or any
-        one-byte-item buffer (see :func:`_wkb_bytes`); the batch narrows
-        nothing.  A list of ``bytes`` or a numpy object array (what
-        ``pandas``/``pyarrow`` hand back for a binary column) both work as
-        they are.  **Byte buffers are first-class, not merely tolerated**: a
-        buffer column costs the same peak a ``bytes`` column does (the table
-        above), so zero-copy ``memoryview`` slices of an Arrow column's value
-        buffer are as cheap an input as ``bytes``.  Cutting those slices out
-        of a pyarrow column correctly is the hard part, and
-        :func:`mortie.arrow.from_wkbs` now does it for you — see Notes.
-    order : int, optional
-        Finest HEALPix order (1-29), shared by every blob.  Default 18.
-    tolerance : float, optional
-        Stop refining a boundary cell at this angular radius in **degrees** —
-        :func:`mortie.morton_coverage_moc`'s ``tolerance``, applied as a
-        **single shared setting**, mutually exclusive with ``max_cells``.
-    max_cells : int, optional
-        Per-blob cell budget, shared by every blob.  A budget below some
-        blob's representable floor is raised for that blob (soft target, as
-        in the scalar path) and one summary warning is emitted.
-    normalize : bool, optional
-        Ring-orientation handling, identical in meaning to
-        :func:`from_wkb`'s ``normalize``.  Default ``True``.
-
-    Returns
-    -------
-    values : numpy.ndarray
-        Every blob's morton MOC words concatenated (``uint64``).
-    out_offsets : numpy.ndarray
-        ``int64`` arrow list offsets into ``values``, length
-        ``len(blobs) + 1``; blob ``i``'s MOC is
-        ``values[out_offsets[i]:out_offsets[i+1]]``.  ``out_offsets[0]`` is
-        always 0 and ``out_offsets[-1]`` is always ``len(values)``.
-
-    Raises
-    ------
-    ValueError
-        Fail-fast, naming the **lowest-index** offending blob (e.g.
-        ``blob 4217: truncated WKB ...``): a malformed or truncated blob, an
-        unclosed polygon ring, an unsupported or empty geometry, a ring with
-        fewer than 3 vertices, or a NaN/infinite coordinate.  **Linear
-        geometry is refused by index** — a LineString cover is one array per
-        line, which has no single-MOC-per-blob spelling; use
-        :func:`from_wkb` for those.  Also for ``order`` outside 1-29, both
-        ``tolerance`` and ``max_cells`` given, or an invalid hex string.
-    TypeError
-        Naming the offending index, for an entry that is neither a string nor
-        a buffer of bytes.
-
-    Notes
-    -----
-    Two ordered gates, as in :func:`mortie.polygons_to_morton_mocs`: the input
-    contract is screened by a serial pre-pass over the whole sequence, then
-    the blobs are parsed and covered.  Each gate reports its own lowest-index
-    offender, so a ``TypeError`` at a high index does surface ahead of a
-    malformed blob at a lower one — the pre-pass is an earlier gate, not a
-    competing one.  The pre-pass **validates without retaining**
-    (``_wkb_bytes(..., materialize=False)``): it applies the identical accept
-    list — an invalid hex string is still caught here, ahead of any parse
-    error — but keeps the entries as they came, so a column in a non-``bytes``
-    spelling is not duplicated for the duration of the call.
-
-    Feeding a **pyarrow** column has a typed entry point of its own —
-    :func:`mortie.arrow.from_wkbs` (issue #163) — and that is what to call:
-    ``marrow.from_wkbs(column, order=...)`` returns exactly this pair.  Do
-    not improvise the extraction.  Four traps sit between a column and its
-    blobs.  **Three are silent** — they yield different, valid-looking data
-    rather than an error: a parquet column reads back as a ``ChunkedArray``,
-    which has no ``.buffers()`` at all; ``slice`` and ``take`` are zero-copy
-    metadata, so a chunk's buffers belong to the *original* array and must be
-    indexed from ``chunk.offset``; and a ``large_binary`` column's offsets are
-    ``int64``, not ``int32``.  The fourth is a **wrong diagnosis**: a null
-    entry spans zero bytes, so it arrives as an empty blob and this function
-    reports it as a truncated geometry rather than as a missing one.  The
-    skin handles all four, and hands this function zero-copy ``memoryview``
-    slices.
-
-    ``from_wkbs(column.to_pylist(), ...)`` is also right on every one of
-    those cases and needs no pyarrow-typed call, but its cost is real: on the
-    555,867-row ATL03 v007 WKB column (290.1 MB of payload) ``to_pylist()``
-    peaks at **~322 MB** of Python ``bytes`` objects — the per-object
-    overhead on top of the payload, plus the list — against **~112 MB** for
-    the skin's views of the same column.
-
-    Warns
-    -----
-    UserWarning
-        If ``max_cells`` is below the minimum needed to represent some blob;
-        the warning reports how many were raised and names the lowest-index
-        one.
-
-    See Also
-    --------
-    from_wkb : the scalar (one blob) form, and the input contract in full.
-
-    Examples
-    --------
-    >>> import mortie                                    # doctest: +SKIP
-    >>> values, off = mortie.from_wkbs(wkb_column, order=8)   # doctest: +SKIP
-    >>> first = values[off[0]:off[1]]   # the first blob's MOC
-    """
-    from . import _rustie
-
-    if tolerance is not None and max_cells is not None:
-        raise ValueError("pass at most one of tolerance / max_cells")
-    # Serial screening pass, in index order, so the lowest-index bad entry is
-    # what a caller sees -- the same fail-fast rule the Rust side applies to
-    # parse/cover failures.  It *validates* rather than coerces: the entries
-    # are kept as they came, so this costs a list of pointers whatever spelling
-    # the column is in, and the byte-producing coercion happens per chunk on
-    # the Rust side, where it is released with the chunk (issue #157).
-    entries = []
-    for i, blob in enumerate(blobs):
-        try:
-            _wkb_bytes(blob, materialize=False)
-        except (TypeError, ValueError) as exc:
-            raise type(exc)(f"blob {i}: {exc}") from exc
-        entries.append(blob)
-    tol_rad = None if tolerance is None else np.radians(float(tolerance))
-    values, out_offsets = _rustie.rust_wkbs_coverage_mocs(
-        entries, _wkb_bytes, order, tol_rad, max_cells, normalize
-    )
-    return np.asarray(values), np.asarray(out_offsets)
-
-
 def from_wkt(text, order=18, moc=False, normalize=True,
              tolerance=None, max_cells=None):
     """Cover a geometry given as WKT (or EWKT) text.
